API proxy for KITSU
